+++ platform/windows-webkit/src/XanhCredentialBridge.h
@@ -10,6 +10,7 @@
 #include <WebKit/WKRetainPtr.h>
 #include <WebKit/WebKit2_C.h>
 #include <functional>
+#include <memory>
 #include <optional>
 
 class XanhCredentialBridge {
@@ -18,7 +19,8 @@ public:
         std::wstring username;
         std::wstring password;
     };
-    using Picker = std::function<std::optional<Credential>(const XanhCredentialBridgePolicy::Request&)>;
+    using PickerCompletion = std::function<void(std::optional<Credential>)>;
+    using Picker = std::function<void(XanhCredentialBridgePolicy::Request, PickerCompletion)>;
     using ForegroundCheck = std::function<bool()>;
 
     XanhCredentialBridge(WKPageConfigurationRef, bool isPrivate, Picker = { }, ForegroundCheck = { });
@@ -35,8 +37,13 @@ public:
     void rendererTerminated(WKPageRef);
 
 private:
+    struct Lifetime;
+    struct PendingRequest;
+
     static void didReceiveScriptMessage(WKScriptMessageRef, WKCompletionListenerRef, const void*);
     void handleScriptMessage(WKScriptMessageRef, WKCompletionListenerRef);
+    void finishPendingRequest(const std::shared_ptr<PendingRequest>&, std::optional<Credential>);
+    void cancelPendingRequest();
 
     WKRetainPtr<WKDictionaryRef> unavailableReply(std::wstring_view requestID) const;
     WKRetainPtr<WKDictionaryRef> credentialReply(std::wstring_view requestID, const Credential&) const;
@@ -44,8 +51,10 @@ private:
     WKRetainPtr<WKUserContentControllerRef> m_controller;
     WKPageRef m_page { nullptr };
     XanhCredentialBridgePolicy::State m_state;
+    XanhCredentialBridgePolicy::AsyncRequestGate m_asyncGate;
     Picker m_picker;
     ForegroundCheck m_foregroundCheck;
-    bool m_requestInFlight { false };
+    std::shared_ptr<Lifetime> m_lifetime;
+    std::shared_ptr<PendingRequest> m_pendingRequest;
     bool m_enabled { false };
 };